+++ kroxylicious-kms-provider-hashicorp-vault/src/test/java/io/kroxylicious/kms/provider/hashicorp/vault/VaultKmsTest.java
@@ -31,7 +31,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class VaultKmsTest {
 
     // address in TEST-NET-1, reserved for use in example specifications and other documents
-    private static final URI NON_ROUTABLE = URI.create("http://192.0.2.1/");
+    private static final URI NON_ROUTABLE = URI.create("http://192.0.2.1/v1/transit");
     private static final byte ARBITRARY_BODY_BYTE = 5;
 
     @Test
@@ -46,8 +46,9 @@ class VaultKmsTest {
     @Test
     void testRequestTimeoutWaitingForHeaders() {
         HttpServer httpServer = delayResponse(Duration.ofSeconds(1), Duration.ZERO);
-        URI address = URI.create("http://" + httpServer.getAddress().getHostName() + ":" + httpServer.getAddress().getPort() + "/");
-        VaultKms kms = new VaultKms(address, "token", Duration.ofMillis(500), null);
+        var addr = httpServer.getAddress();
+        var uri = URI.create("http://" + addr.getHostName() + ":" + addr.getPort() + "/v1/transit");
+        VaultKms kms = new VaultKms(uri, "token", Duration.ofMillis(500), null);
         CompletableFuture<String> alias = kms.resolveAlias("alias");
         assertThat(alias).failsWithin(Duration.ofSeconds(2))
                 .withThrowableOfType(ExecutionException.class)
@@ -55,11 +56,9 @@ class VaultKmsTest {
         httpServer.stop(0);
     }
 
-    static Stream<Arguments> legalEnginePaths() {
+    static Stream<Arguments> acceptableVaultTransitEnginePaths() {
         var uri = URI.create("https://localhost:1234");
         return Stream.of(
-                Arguments.of("no path", uri, "/v1/transit/"),
-                Arguments.of("slash only", uri.resolve("/"), "/v1/transit/"),
                 Arguments.of("basic", uri.resolve("v1/transit"), "/v1/transit/"),
                 Arguments.of("trailing slash", uri.resolve("v1/transit/"), "/v1/transit/"),
                 Arguments.of("single namespace", uri.resolve("v1/ns1/transit"), "/v1/ns1/transit/"),
@@ -68,27 +67,29 @@ class VaultKmsTest {
     }
 
     @ParameterizedTest(name = "{0}")
-    @MethodSource
-    void legalEnginePaths(String name, URI uri, String expected) {
+    @MethodSource("acceptableVaultTransitEnginePaths")
+    void acceptsVaultTransitEnginePaths(String name, URI uri, String expected) {
         var kms = new VaultKms(uri, "token", Duration.ofMillis(500), null);
-        assertThat(kms.getEngineUri())
+        assertThat(kms.getVaultTransitEngineUri())
                 .extracting(URI::getPath)
                 .isEqualTo(expected);
 
     }
 
-    static Stream<Arguments> illegalEnginePaths() {
+    static Stream<Arguments> unacceptableVaultTransitEnginePaths() {
         var uri = URI.create("https://localhost:1234");
         return Stream.of(
-                Arguments.of("unrecognized version", uri.resolve("v999/transit")),
+                Arguments.of("no path", uri),
+                Arguments.of("missing path", uri.resolve("/")),
+                Arguments.of("unrecognized API version", uri.resolve("v999/transit")),
                 Arguments.of("missing engine", uri.resolve("v1")),
                 Arguments.of("empty engine", uri.resolve("v1/")));
     }
 
     @ParameterizedTest(name = "{0}")
-    @MethodSource
-    void illegalEnginePaths(String name, URI uri) {
-        assertThatThrownBy(() -> new VaultKms(uri, "token", Duration.ofMillis(500), null))
+    @MethodSource("unacceptableVaultTransitEnginePaths")
+    void detectsUnacceptableVaultTransitEnginePaths(String name, URI uri) {
+        assertThatThrownBy(() -> new VaultKms(uri, "token", Duration.ZERO, null))
                 .isInstanceOf(IllegalArgumentException.class);
 
     }
@@ -97,8 +98,9 @@ class VaultKmsTest {
     @Test
     void testRequestTimeoutWaitingForBody() {
         HttpServer httpServer = delayResponse(Duration.ZERO, Duration.ofSeconds(1));
-        URI address = URI.create("http://" + httpServer.getAddress().getHostName() + ":" + httpServer.getAddress().getPort() + "/");
-        VaultKms kms = new VaultKms(address, "token", Duration.ofMillis(500), null);
+        var addr = httpServer.getAddress();
+        URI uri = URI.create("http://" + addr.getHostName() + ":" + httpServer.getAddress().getPort() + "/v1/transit");
+        VaultKms kms = new VaultKms(uri, "token", Duration.ofMillis(500), null);
         CompletableFuture<String> alias = kms.resolveAlias("alias");
         assertThat(alias).failsWithin(Duration.ofSeconds(2))
                 .withThrowableOfType(ExecutionException.class)
@@ -125,6 +127,7 @@ class VaultKmsTest {
         }
     }
 
+    @SuppressWarnings("java:S2925")
     private static void sleep(Duration delayHeaders) {
         try {
             Thread.sleep(delayHeaders.toMillis());