+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Please enumerate **all user-facing** changes using format `<githib issue/pr numb
   adjust your dependencies as your adopt this release.
 * `io.kroxylicious:kroxylicious-filter-test-support` now contains RecordTestUtils for creating example `Record`, `RecordBatch` and `MemoryRecords`. It also contains 
   assertj assertions for those same classes to enable us to write fluent assertions, accessible via `io.kroxylicious.test.assertj.KafkaAssertions`.
+* The configuration for VaultKMS service has changed.  Instead of the `vaultUrl` config key, the provider now
+  requires `vaultTransitEngineUrl`.  This must provide the complete path to the Transit Engine on the
+  HashiCorp Vault instance (e.g. https://myvault:8200/v1/transit or https://myvault:8200/v1/mynamespace/transit).  
 
 ## 0.4.1